Define faqs in the frontmatter for case studies.
